Release/1.0.0 (#20)
* test: MatchRepositoryTest. CRUD tests

* test: fixed names of methods in PlayerRepositoryTest. Added tests for custom methods of MatchRepository

* test: added new methods to MatchRepository

* Integration tests for RoundRepository (#2)

- tested CRUD methods
- tested bidirectional relationships with match
- tested custom methods of RoundRepository

* test: Integration tests for Tournament Repository

- Modified one Tournament's constructor to chain it to the other ones
- Added integration tests for TournamentRepository

* Bugfix/controllers should return 404 status when source is not found (#4)

* Fixed incorrect responses for TournamentController, PlayerController and RoundController. 
* The call to MatchService from RoundController has been redirected through a facade method in RoundService

* Feature/logging: added logger system (#5)

* feat: added logger info messages to every endpoint. refactor: now deleteTournament() from TournamentController hasn't a catch for TournamentException, since the exception is already handled at exceptionHandler

* feature/logging: feat: added fine and finer logging for services.

feat: created Log formatters to ensure a unified log format

feat: added a CustomRequestContextFilter

refactor: logs at controllers now use the new log formatters

* feature/logging: feat: a log has been added for the cases of unexpected RuntimeExceptions

* Doc/readme update (#6)

* doc: Update README.md
* doc: added CONTRIBUTING.md
* refactor: removed unused imports

* Update README.md (#7)

* Docs/readme update endpoints (#9)

* Create API-ENDPOINTS.md

* Update README.md

- added link to endpoints section to Readme.md

* Added line breaks to API-ENDPOINTS.md

* docs: fixed main branch's name on Contributing.md

* doc: added new md files for every endpoint exposed on TournamentController

fix: the endpoint for delete tournament no longer returns a message beyond the status 204 No Content

* doc: added .md doc files for every endpoint in RoundController and MatchController

* doc: added a new md file for every playerController endpoint

fix: changed ok response from deletePlayer, to no content

* Docs/requirements update (#10)

doc: Updated README.md:
- doc: added link to requirements section
- doc: now Available endpoints has changed to Endpoints
- doc: the titles are links to their own sections

doc: Create API-REQUIREMENTS.md

* Docs/test (#11)

* doc: added link to tests section to README.md

* doc: Create API-TESTS.md

* Docs/persistence (#12)

- added section: Persistence and Data Storage to README.md
- created API-PERSISTENCE.md

* finished section structure (#13)

* doc: Readme.md now has a link to project structure section

* doc: added md file for the section project structure

* doc: added Packages subsection to Project structure section

* Docs/security (#14)

* doc: added API-SECURITY.md

* doc: added security section to README.md

* Docs/versioning (#15)

* doc: Create API-VERSIONING.md

* doc: added Project status and Versioning sections to README.md

* doc: CONTRIBUTING.md has been translated to english

* Update CONTRIBUTING.md

* Docs/installation (#16)

* doc: readme.md now includes sections for installation and ci cd

* doc: Create API-CICD.md

* Create API-INSTALLATION.md

* Update API-INSTALLATION.md (#17)

* Update API-INSTALLATION.md (#18)

* Create LICENSE

* chore(release): prepare version 1.0.0

* updated jenkinsfiles and pom version

* feat: Jenkinsfile.build now allows to identify the branch that is going to be built

* feat: now deployment jenkinsfiles now allow to choose the branch to be deployed

* doc: CICD documentation updated

+++ src/main/java/com/git/luisdeveloper/wargames_tournament/service/impl/PlayerServiceImpl.java
@@ -1,6 +1,7 @@
 package com.git.luisdeveloper.wargames_tournament.service.impl;
 
 import java.util.List;
+import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,7 @@ import com.git.luisdeveloper.wargames_tournament.dto.UpdatePasswordDTO;
 import com.git.luisdeveloper.wargames_tournament.dto.UpdatePersonalDataDTO;
 import com.git.luisdeveloper.wargames_tournament.exception.InvalidCredentialsException;
 import com.git.luisdeveloper.wargames_tournament.exception.PlayerNotFoundException;
+import com.git.luisdeveloper.wargames_tournament.logging.ServiceLogFormatter;
 import com.git.luisdeveloper.wargames_tournament.mappers.PlayerMapper;
 import com.git.luisdeveloper.wargames_tournament.repository.PlayerRepository;
 import com.git.luisdeveloper.wargames_tournament.service.PlayerService;
@@ -18,6 +20,11 @@ import jakarta.transaction.Transactional;
 
 @Service
 public class PlayerServiceImpl implements PlayerService {
+
+	@Autowired
+	private ServiceLogFormatter formatter;
+
+	private static Logger logger = Logger.getLogger(PlayerServiceImpl.class.getName());
 
 	@Autowired
 	private PlayerRepository repository;
@@ -33,31 +40,52 @@ public class PlayerServiceImpl implements PlayerService {
 
 	@Override
 	@Transactional
-	public void updatePlayer(UpdatePasswordDTO dto) throws InvalidCredentialsException {
+	public void updatePlayer(UpdatePasswordDTO dto) throws InvalidCredentialsException, PlayerNotFoundException {
+		logger.fine(formatter.request("updating player with id: " + dto.playerId()));
+		if (!repository.existsById(dto.playerId())) {
+			throw new PlayerNotFoundException();
+		}
 		int updatedEntries = repository.updatePassword(dto.playerId(), dto.oldPassword(), dto.newPassword());
 		if (updatedEntries == 0)
 			throw new InvalidCredentialsException();
+		logger.fine(formatter.success("updating player with id: " + dto.playerId()));
 
 	}
 
 	@Override
 	public List<PlayerRankingDTO> getRanking() {
-		return repository.findAll().stream().map(x -> PlayerMapper.toPlayerRankingDTO(x)).sorted().toList();
+		logger.fine(formatter.request("Get ranking"));
+		var ranking = repository.findAll().stream().map(x -> PlayerMapper.toPlayerRankingDTO(x)).sorted().toList();
+		logger.fine(formatter.success("Get ranking"));
+		return ranking;
 	}
 
 	@Override
 	public int updatePlayerPoints(Long id, int points) {
-		return repository.updatePlayerPoints(id, points);
+		logger.fine(formatter.request("Updating points of player with id: " + id));
+		var numUpdatedPlayers = repository.updatePlayerPoints(id, points);
+		logger.fine(formatter.success("Updating points of player with id: " + id));
+		return numUpdatedPlayers;
 	}
 
 	@Override
 	public List<PlayerRankingDTO> getPlayers() {
-		return repository.findAll().stream().map(x -> PlayerMapper.toPlayerRankingDTO(x)).toList();
+		logger.fine(formatter.request("Getting ranking"));
+		var ranking = repository.findAll().stream().map(x -> PlayerMapper.toPlayerRankingDTO(x)).toList();
+		logger.fine(formatter.success("Getting ranking"));
+		return ranking;
 	}
 
 	@Override
 	public void deletePlayer(Long id) throws PlayerNotFoundException {
-		repository.deleteById(id);
+		logger.fine(formatter.request("Deleting player with id: " + id));
+		if (repository.existsById(id)) {
+			repository.deleteById(id);
+			logger.fine(formatter.success("Deleting player with id: " + id));
+		}else {
+			throw new PlayerNotFoundException();
+		}
+		
 	}
 
 }
